Fix en tout genre.
- Oubli d'un " ; " dans un fichier header corriger.
- size_t remplacer par uint_32t car on compile vers une architecture
  cible 32bits, or size_t remplace par uint_64t car processeur x86_64.
- Le linker ne linkais pas les fichier printk.o et utils.o, cela a ete
  ajouter et corriger.
- Ajouts de regles makefile run et run_debug.
- Correction d'erreur de printk qui ne changais jamais de caracteres et
  affichais donc en boule uniquement le premier sur la longeure de la
chaine.
- Le linker se plaignais que les fichiers .o etaient compiler pour du
  x86_64 ce qui faisait planter le linker, cela a ete corriger, mais
aucune idee de ce qui as causer le probleme en premier lieux.

+++ C/src/utils.c
@@ -1,8 +1,8 @@
 #include "../headers/header.h"
 
-size_t ft_strlen(const char* str)
+uint32_t ft_strlen(const char* str)
 {
-	size_t res = 0;
+	uint32_t res = 0;
 
 	while (*(str++))
 	    res++;
